+++ account-kit/smart-contracts/src/ma-v2/client/client.test.ts
@@ -59,11 +59,16 @@ import {
   type ContractFunctionName,
   type TestActions,
 } from "viem";
-import { entryPoint07Abi } from "viem/account-abstraction";
+import {
+  createWebAuthnCredential,
+  entryPoint07Abi,
+  type ToWebAuthnAccountParameters,
+} from "viem/account-abstraction";
 import { generatePrivateKey, privateKeyToAccount } from "viem/accounts";
 import { setBalance } from "viem/actions";
 import { local070Instance } from "~test/instances.js";
 import { paymaster070 } from "~test/paymaster/paymaster070.js";
+import { SoftWebauthnDevice } from "~test/webauthn.js";
 import {
   packAccountGasLimits,
   packPaymasterData,
@@ -131,6 +136,64 @@ describe("MA v2 Tests", async () => {
       startingAddressBalance + sendAmount,
     );
   });
+
+  it("sends a simple UO with webauthn account", async () => {
+    const { provider } = await givenWebAuthnProvider();
+
+    await setBalance(instance.getClient(), {
+      address: provider.getAddress(),
+      value: parseEther("2"),
+    });
+
+    const startingAddressBalance = await getTargetBalance();
+
+    const result = await provider.sendUserOperation({
+      uo: {
+        target: target,
+        value: sendAmount,
+        data: "0x",
+      },
+    });
+
+    await provider.waitForUserOperationTransaction(result).catch(async () => {
+      const dropAndReplaceResult = await provider.dropAndReplaceUserOperation({
+        uoToDrop: result.request,
+      });
+      await provider.waitForUserOperationTransaction(dropAndReplaceResult);
+    });
+
+    await expect(getTargetBalance()).resolves.toEqual(
+      startingAddressBalance + sendAmount,
+    );
+  });
+
+  it.fails(
+    "successfully sign + validate a message, for WebAuthn account",
+    async () => {
+      const { provider } = await givenWebAuthnProvider();
+
+      await setBalance(instance.getClient(), {
+        address: provider.getAddress(),
+        value: parseEther("2"),
+      });
+
+      const message = "0xdeadbeef";
+
+      let signature = await provider.signMessage({ message });
+
+      const publicClient = instance.getClient().extend(publicActions);
+
+      // TODO: should be using verifyTypedData here
+      const isValid = await publicClient.verifyMessage({
+        // TODO: this is gonna fail until the message can be formatted since the actual message is EIP-712
+        message,
+        address: provider.getAddress(),
+        signature,
+      });
+
+      expect(isValid).toBe(true);
+    },
+  );
 
   it("successfully sign + validate a message, for native and single signer validation", async () => {
     const provider = (await givenConnectedProvider({ signer })).extend(
@@ -1815,6 +1878,60 @@ describe("MA v2 Tests", async () => {
   });
 
   let salt = 1n;
+
+  const givenConnectedWebauthnProvider = async ({
+    signerEntity,
+    accountAddress,
+    paymasterMiddleware,
+    credential,
+    getFn,
+    rpId,
+  }: {
+    signerEntity?: SignerEntity;
+    accountAddress?: `0x${string}`;
+    paymasterMiddleware?: "alchemyGasAndPaymasterAndData" | "erc7677";
+    credential: ToWebAuthnAccountParameters["credential"];
+    getFn?: ToWebAuthnAccountParameters["getFn"];
+    rpId?: ToWebAuthnAccountParameters["rpId"];
+  }) =>
+    createModularAccountV2Client({
+      chain: instance.chain,
+      accountAddress,
+      signerEntity,
+      credential,
+      getFn,
+      rpId,
+      mode: "webauthn",
+      transport: custom(instance.getClient()),
+      ...(paymasterMiddleware === "alchemyGasAndPaymasterAndData"
+        ? alchemyGasAndPaymasterAndDataMiddleware({
+            policyId: "FAKE_POLICY_ID",
+            // @ts-ignore (expects an alchemy transport, but we're using a custom transport for mocking)
+            transport: custom(instance.getClient()),
+          })
+        : paymasterMiddleware === "erc7677"
+          ? erc7677Middleware()
+          : {}),
+      salt: salt++,
+    });
+
+  const givenWebAuthnProvider = async () => {
+    const webauthnDevice = new SoftWebauthnDevice();
+
+    const credential = await createWebAuthnCredential({
+      rp: { id: "localhost", name: "localhost" },
+      createFn: (opts) => webauthnDevice.create(opts, "localhost"),
+      user: { name: "test", displayName: "test" },
+    });
+
+    const provider = await givenConnectedWebauthnProvider({
+      credential,
+      getFn: (opts) => webauthnDevice.get(opts, "localhost"),
+      rpId: "localhost",
+    });
+
+    return { provider, credential };
+  };
 
   const givenConnectedProvider = async ({
     signer,